rpc/jsonrpc: reject an eth_getTxWitness index that overflows int
hexutil.Uint is 64-bit, so int(txIndex) wrapped negative above math.MaxInt64 and
the bound never fired: the call returned a full block witness with no error.

+++ rpc/jsonrpc/eth_call.go
@@ -705,7 +705,7 @@ func (api *BaseAPI) getWitness(ctx context.Context, db kv.TemporalRoDB, blockNrO
 		return nil, nil
 	}
 
-	if !fullBlock && int(txIndex) >= len(block.Transactions()) {
+	if !fullBlock && uint64(txIndex) >= uint64(len(block.Transactions())) {
 		return nil, fmt.Errorf("transaction index out of bounds: %d", txIndex)
 	}
